use fnv instead of incorrect md5
The homegrown Sum32 was wrong and md5 is slow, anyhow. fnv isn't
stellar, but it's predictable and folks can swap it out in their own
implementations.

The test vectors get some mild changes to account for the new hash, and
we litter some t.Helper calls to clarify test failures.

Fixes #11